Refac media/pdf thumbnailers Now MediaThumbnailerWorker delegates all the work to MediaThumbnailer service
